Change how coverpoints and covergroups are instanced (#49)
Coverpoints and covergroups are now instanced by using _init(), called
by the add_* functions.
CoverTop has been added for the top of coverage. 
default names, descriptions and motivations (along with tier and tags)
can now be part of the coverpoint/covergroup code. And can be overridden
when adding to a covergroup.